requirements.txt
streamlit
requests
openai
pandas
Pillow